Adds a SpeedController and Makefile.
This SpeedController is intended to regulate motor speed for all sources
of input to the VESC (Nunchuk, PPM, etc.). It is manages all aspects of
the motor drive given a simple API.

This is likely to change very much as work is continued, but has been
tested with ~10 miles of riding so far and is quite stable. To be ported
from the existing nunchuk is support for cruise control and traction
control.