Adds an Animation and supporting Interpolators.
An Animation invokes a callback for each frame of an animation and
provides a floating point value interpolated between two values (start
and stop). The values provided are obtained from an Interpolator
function which can implement any variety of time steps including Linear,
Step and Cubic.

The API is flexible and the inteprolators can be extended to support any
of the CSS interpolators (and more) if needed.