hw_rd2: Added macro RADINN_SUB_FW_VERS.
Added macro for tracking Radinn sub-revisions on top of the current v5 baseline.
For this purpose the bldc fw version field "FW_TEST_VERSION_NUBMER" is
high-jacked (used for internal sub-releases before official release so should be
safe). And also, adjusted the naming of the hw_rd2.h inclusion-guard and the
hw_rd2_get_temp() function.

Current Radinn sub-revision is '0'.