Balance-app: Introduce optional "Reverse-Stop" behavior
This feature, when enabled, allows riders to stop their vehicle by simply riding
in reverse. A minimum forward speed threshold can be configured (via ERPM).

Once the minimum speed threshold is exceeded for the first time the reverse stop
feature will be "armed". Afterwards when the rider comes to a standstill and goes
the slightest bit in reverse the board will stop (triggering a REVERSE_FAULT).

This allows new riders to stop without learning how to heel-lift or
having to jump off.

Signed-off-by: Dado Mista <dadomista@gmail.com>