Balance App: Trigger DUTY fault in MC Interface
When a balance duty fault occurs (duty cycle > configured limit)
the consequence is a nosedive in 99% of cases!
Furthermore it is usually difficult to prove that this happened
even when logs are available.

Propagating this fault over to mc_interface allows duty cycle related
problems to show up in logs and get reported by apps.

Signed-off-by: Dado Mista <dadomista@gmail.com>