[Trig][FixedPoint] Create trigonometry for THUMB-2 fixed-point calculation
Uses fixed-point characteristics to greatly reduce computation time, as
well as improve accuracy.

Exact numbers are hard to come by, but it seems there's a 3-5x speed up as
well as a global 50% improvement in error.

The speed gains are due to the SMLAWx instruction, which allows a
signed multiply and accumulate, combined with a bitshift.

The fixed-point conversion takes advantage of a mapping from 2PI radians to
2[units]. The interested reader should inspect the code comments for a
better description.