[Math][Optimization] Provide a CPU-efficient boxcar average
This is the classic CPU-efficicient implementation of a moving average
filter. Rather than recompute the entire sum for i=k..k+N, it exploites the
fact that the sum from i=k-1..k+N-1 is known from the previous iteration.
Thus, all that remains is to do is remove val[k-1] and add val[k+N].