Add regen cutoff to prevent overcharging the battery during braking
This change helps ensure that the battery remains in good condition, prolongs its overall lifespan and avoids hazardous scenarios of overcharge.
A soft cutoff for the regenerative braking is implemented, in which start and end thresholds should be set in VESC Tool. The end threshold should be set such that the cells don't exceed 4.2v in a typical Li-Ion battery.

Signed-off-by: Marcos Chaparro <marcos@lunacycle.com>