Perform full offset calibration when running FOC detection
Undriven offsets are not measured on startup as the motor might be spinning (reboot from crash, etc) So the only time undriven was measured was manually in VESC Tool.
This runs the full offset calibration before detecting the motor to improve detection and make sure the undriven offsets are measured.