hwconf/*: disable DAC buffer to increase accuracy
According to the STM32F4 datasheet, the ADC is equipped with a buffer
for being able to drive external loads directly. This buffer negatively
impacts accuracy, though. The offset can be as high as 0.2V.

Thus, disable the buffer by setting the BOFFx flag.

This was verified on a STM32 nucleo64 F446RE by reading the DAC voltage
back via ADC and comparing results with and without BOFFx=1. The output
voltage was verified (calibrated DMM BM789) to be correct with BOFFx=1.

Reference:
STM DS8626 STM32F40xxx Rev 9
 - 5.3.25 DAC electrical characteristics
 - Figure 53. 12-bit buffered non-buffered DAC

Signed-off-by: Michael Niewöhner <foss@mniewoehner.de>