Add support for user git commit hash metadata
This is intended for projects with their own git repo that use bldc as a
dependency. They can then set `USER_GIT_COMMIT_HASH` to their commit
hash when running make, and it will show up when running `fw_info` in
the terminal.

Also renamed `fwinfo` to `fw_info` for consistency with other terminal
commands.