Fix windows build issue: Build gets stuck if sh.exe is in PATH
On Windows the $(shell ...) function uses sh.exe (BASH) if it is in PATH, otherwise it uses cmd.exe.
If SH is used then the "/C" option is interred as the C:\ drive and not the CMD option which tells the CMD to run the command and exit.
Since "/C" is interpreted the wrong way CMD doesn't exit and the MAKE process gets stuck (doesn't return an error or continue).
But if CMD is used then "/C" is interred correctly and the build continues.
sh.exe can get added to PATH when installing git, but it depends on the selected options.
The solution is to just use POWERSHELL to create the directory.